refactor: 구조 정리
