+++ app/src/main/java/com/issueissyu/fe/data/remote/dto/community/CommunityMappers.kt
@@ -1,5 +1,8 @@
 package com.issueissyu.fe.data.remote.dto.community
 
+import com.issueissyu.fe.data.remote.dto.response.CommunityDetailResponse
+import com.issueissyu.fe.data.remote.dto.response.CommunityFeedItemResponse
+import com.issueissyu.fe.data.remote.dto.response.CommunityFeedResponse
 import com.issueissyu.fe.domain.model.community.CommunityDetail
 import com.issueissyu.fe.domain.model.community.CommunityFeed
 import com.issueissyu.fe.domain.model.community.CommunityFeedItem
