+++ app/src/main/java/com/issueissyu/fe/data/remote/dto/response/CommunityDetailResponse.kt
@@ -1,4 +1,4 @@
-package com.issueissyu.fe.data.remote.dto.community
+package com.issueissyu.fe.data.remote.dto.response
 
 data class CommunityDetailResponse(
     val item: CommunityFeedItemResponse?,
