+++ app/src/main/java/com/issueissyu/fe/data/remote/dto/response/CommunityFeedResponse.kt
@@ -1,4 +1,4 @@
-package com.issueissyu.fe.data.remote.dto.community
+package com.issueissyu.fe.data.remote.dto.response
 
 data class CommunityFeedResponse(
     val region: String?,
