+++ app/src/main/java/com/issueissyu/fe/di/RepositoryModule.kt
@@ -1,12 +1,12 @@
 package com.issueissyu.fe.di
 
 import com.issueissyu.fe.data.repository.AuthRepositoryImpl
-import com.issueissyu.fe.data.repository.DefaultIssueRepository
-import com.issueissyu.fe.data.repository.IssueRepository
-import com.issueissyu.fe.data.repository.UserRepository
+import com.issueissyu.fe.domain.repository.DefaultIssueRepository
+import com.issueissyu.fe.domain.repository.IssueRepository
+import com.issueissyu.fe.domain.repository.UserRepository
 import com.issueissyu.fe.data.repository.UserRepositoryImpl
 import com.issueissyu.fe.data.repository.LocationRepositoryImpl
-import com.issueissyu.fe.data.repository.PinRepository
+import com.issueissyu.fe.domain.repository.PinRepository
 import com.issueissyu.fe.data.repository.PinRepositoryImpl
 
 import com.issueissyu.fe.domain.repository.AuthRepository
