+++ app/src/main/java/com/issueissyu/fe/domain/model/MapPinModels.kt
@@ -1,4 +1,4 @@
-package com.issueissyu.fe.data.model
+package com.issueissyu.fe.domain.model
 
 data class MapPinMarker(
     val pinId: String,
