+++ app/src/main/java/com/issueissyu/fe/domain/model/auth/AuthUser.kt
@@ -1,4 +1,4 @@
-package com.issueissyu.fe.domain.model
+package com.issueissyu.fe.domain.model.auth
 
 data class AuthUser(
     val uuid: String,
