+++ app/src/main/java/com/issueissyu/fe/ui/IssueissyuApp.kt
@@ -11,8 +11,7 @@ import com.issueissyu.fe.ui.navigation.AppNavGraph
 
 private fun shouldShowBottomBar(route: String?): Boolean {
     if (route == null) return false
-    return route == AppDestinations.HOME_ROUTE ||
-        route == AppDestinations.COLLECTION_ROUTE ||
+    return route == AppDestinations.COLLECTION_ROUTE ||
         route == AppDestinations.TOWN_ROUTE ||
         route == AppDestinations.COMMUNITY_ROUTE ||
         route == AppDestinations.MYPAGE_ROUTE ||
