+++ app/src/main/java/com/issueissyu/fe/ui/screens/map/MapMarkerStyleMapper.kt
@@ -1,7 +1,7 @@
 package com.issueissyu.fe.ui.screens.map
 
 import com.issueissyu.fe.R
-import com.issueissyu.fe.data.model.PinCategory
+import com.issueissyu.fe.domain.model.PinCategory
 
 fun PinCategory.toMarkerIconRes(): Int {
     return when (this) {
