+++ app/src/main/java/com/issueissyu/fe/ui/screens/onboarding/LoginEffect.kt
@@ -1,4 +1,4 @@
-package com.issueissyu.fe.ui.viewmodels
+package com.issueissyu.fe.ui.screens.onboarding
 
 //로그인 성공 후 화면
 sealed interface LoginEffect {
