+++ app/src/main/java/com/issueissyu/fe/ui/screens/pindetail/PinDetailViewModel.kt
@@ -2,8 +2,8 @@ package com.issueissyu.fe.ui.screens.pindetail
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.issueissyu.fe.data.model.Pin
-import com.issueissyu.fe.data.repository.PinRepository
+import com.issueissyu.fe.domain.model.Pin
+import com.issueissyu.fe.domain.repository.PinRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
