+++ app/src/main/java/com/issueissyu/fe/ui/screens/pindetail/PinHomeTab.kt
@@ -38,11 +38,11 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import coil.compose.AsyncImage
 import com.issueissyu.fe.R
-import com.issueissyu.fe.data.model.CommunicationPinDetail
-import com.issueissyu.fe.data.model.IssuePinDetail
-import com.issueissyu.fe.data.model.Pin
-import com.issueissyu.fe.data.model.ResolutionStatus
-import com.issueissyu.fe.data.model.canEditBy
+import com.issueissyu.fe.domain.model.CommunicationPinDetail
+import com.issueissyu.fe.domain.model.IssuePinDetail
+import com.issueissyu.fe.domain.model.Pin
+import com.issueissyu.fe.domain.model.ResolutionStatus
+import com.issueissyu.fe.domain.model.canEditBy
 import com.issueissyu.fe.data.sample.PinSamples
 import com.issueissyu.fe.ui.theme.BrandColor
 import com.issueissyu.fe.ui.theme.Gray_3
