--- app/src/main/java/com/issueissyu/fe/ui/viewmodels/HomeViewModel.kt
@@ -1,20 +1,0 @@
-package com.issueissyu.fe.ui.viewmodels
-
-import androidx.compose.runtime.State
-import androidx.compose.runtime.mutableStateOf
-import androidx.lifecycle.ViewModel
-import dagger.hilt.android.lifecycle.HiltViewModel
-import javax.inject.Inject
-
-data class HomeUiState(
-    val message: String = "이슈 있슈 FE - 초기 예시 화면입니다"
-)
-
-@HiltViewModel
-class HomeViewModel @Inject constructor() : ViewModel() {
-
-    private val _uiState = mutableStateOf(HomeUiState())
-    val uiState: State<HomeUiState> = _uiState
-
-    // Add more logic here as needed
-}
